cloud_functions/extractors/constant-contact-sync/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.*
google-cloud-secret-manager==2.*
requests==2.*